добавил добавление резервации

+++ db/models.py
@@ -45,4 +45,4 @@ class Order(models.Model):
     room = models.ForeignKey(Room, on_delete=models.CASCADE)
     check_in_date = models.DateField()
     check_out_date = models.DateField()
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
